+++ packages/synapse-interface/slices/transactions/updater.tsx
@@ -171,17 +171,17 @@ export default function Updater(): null {
         ?.filter(
           (recentTx) =>
             (userPendingTransactions &&
-              userPendingTransactions.some(
+              userPendingTransactions?.some(
                 (pendingTx: BridgeTransaction) =>
                   pendingTx.fromInfo.txnHash === recentTx.transactionHash
               )) ||
             (userHistoricalTransactions &&
-              userHistoricalTransactions.some(
+              userHistoricalTransactions?.some(
                 (historicalTx: BridgeTransaction) =>
                   historicalTx.fromInfo.txnHash === recentTx.transactionHash
               )) ||
             (fallbackQueryPendingTransactions &&
-              fallbackQueryPendingTransactions.some(
+              fallbackQueryPendingTransactions?.some(
                 (pendingTx: BridgeTransaction) =>
                   pendingTx.fromInfo.txnHash === recentTx.transactionHash
               ))
@@ -214,7 +214,7 @@ export default function Updater(): null {
     if (hasUserPendingTransactions) {
       userPendingTransactions.forEach(
         (pendingTransaction: BridgeTransaction) => {
-          const isStored: boolean = pendingAwaitingCompletionTransactions.some(
+          const isStored: boolean = pendingAwaitingCompletionTransactions?.some(
             (storedTransaction: BridgeTransaction) =>
               storedTransaction.kappa === pendingTransaction.kappa
           )
@@ -243,7 +243,7 @@ export default function Updater(): null {
       const mostRecentHistoricalTransaction: BridgeTransaction =
         userHistoricalTransactions[0]
 
-      const isTransactionAlreadySeen = seenHistoricalTransactions.some(
+      const isTransactionAlreadySeen = seenHistoricalTransactions?.some(
         (transaction: BridgeTransaction) =>
           transaction === (mostRecentHistoricalTransaction as BridgeTransaction)
       )
@@ -257,12 +257,12 @@ export default function Updater(): null {
       pendingAwaitingCompletionTransactions.forEach(
         (pendingTransaction: BridgeTransaction) => {
           const isCompleted: boolean =
-            userHistoricalTransactions.some(
+            userHistoricalTransactions?.some(
               (historicalTransaction: BridgeTransaction) => {
                 return historicalTransaction.kappa === pendingTransaction.kappa
               }
             ) ||
-            fallbackQueryHistoricalTransactions.some(
+            fallbackQueryHistoricalTransactions?.some(
               (historicalTransaction: BridgeTransaction) => {
                 return historicalTransaction.kappa === pendingTransaction.kappa
               }
@@ -287,7 +287,7 @@ export default function Updater(): null {
       pendingBridgeTransactions.forEach(
         (pendingBridgeTransaction: PendingBridgeTransaction) => {
           const isCompleted: boolean =
-            userHistoricalTransactions.some(
+            userHistoricalTransactions?.some(
               (historicalTransaction: BridgeTransaction) => {
                 return (
                   historicalTransaction.fromInfo.txnHash ===
@@ -295,7 +295,7 @@ export default function Updater(): null {
                 )
               }
             ) ||
-            fallbackQueryHistoricalTransactions.some(
+            fallbackQueryHistoricalTransactions?.some(
               (historicalTransaction: BridgeTransaction) => {
                 return (
                   historicalTransaction.fromInfo.txnHash ===
@@ -326,7 +326,7 @@ export default function Updater(): null {
       const mostRecentFallbackHistoricalTransaction: BridgeTransaction =
         fallbackQueryHistoricalTransactions[0]
 
-      const isTransactionAlreadySeen = seenHistoricalTransactions.some(
+      const isTransactionAlreadySeen = seenHistoricalTransactions?.some(
         (transaction: BridgeTransaction) =>
           transaction.kappa ===
           (mostRecentFallbackHistoricalTransaction.kappa as BridgeTransaction)
@@ -351,11 +351,11 @@ export default function Updater(): null {
         const { fromInfo, toInfo, kappa } = transaction
 
         const alreadyMovedToHistorical: boolean =
-          fallbackQueryHistoricalTransactions.some(
+          fallbackQueryHistoricalTransactions?.some(
             (historicalTransaction: BridgeTransaction) =>
               historicalTransaction !== transaction
           ) ||
-          userHistoricalTransactions.some(
+          userHistoricalTransactions?.some(
             (historicalTransaction: BridgeTransaction) =>
               historicalTransaction !== transaction
           )
@@ -385,7 +385,7 @@ export default function Updater(): null {
       fallbackQueryHistoricalTransactions.forEach(
         (fallbackTransaction: BridgeTransaction) => {
           if (
-            userHistoricalTransactions.some(
+            userHistoricalTransactions?.some(
               (historicalTransaction: BridgeTransaction) =>
                 historicalTransaction.kappa === fallbackTransaction.kappa
             )
@@ -408,7 +408,7 @@ export default function Updater(): null {
     if (checkTransactionsExist(fallbackQueryHistoricalTransactions)) {
       fallbackQueryHistoricalTransactions.forEach(
         (fallbackHistoricalTransaction: BridgeTransaction) => {
-          const matched: boolean = fallbackQueryPendingTransactions.some(
+          const matched: boolean = fallbackQueryPendingTransactions?.some(
             (pendingTransaction: BridgeTransaction) =>
               pendingTransaction.kappa === fallbackHistoricalTransaction.kappa
           )
@@ -426,7 +426,7 @@ export default function Updater(): null {
     if (checkTransactionsExist(fallbackQueryPendingTransactions)) {
       fallbackQueryPendingTransactions.forEach(
         (fallbackPendingTransaction: BridgeTransaction) => {
-          const matched: boolean = userHistoricalTransactions.some(
+          const matched: boolean = userHistoricalTransactions?.some(
             (pendingTransaction: BridgeTransaction) =>
               pendingTransaction.kappa === fallbackPendingTransaction.kappa
           )